Mid-way through file cleanup.

+++ Flow/Source/Flow/Framework/Utils/Profiling.cpp
@@ -84,7 +84,7 @@ void BenchmarkTimer::Stop()
 	long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartPoint).time_since_epoch().count();
 	long long end = std::chrono::time_point_cast<std::chrono::microseconds>(EndPoint).time_since_epoch().count();
 
-	uint32_t ThreadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+	uint32_t ThreadID = (uint32_t)std::hash<std::thread::id>{}(std::this_thread::get_id());
 	Instrumentor::Get().WriteProfile({ m_TimerName, start, end, ThreadID });
 
 	m_Stopped = true;
